If running git clean after reset then retry
* devstack-vm-gate-wrap.sh: If a git clean can not remove some
files, for example because a previous command has not freed its open
file handles, it will return nonzero. If this happens, wait a moment
and try again in case this is only a briefly transient problem.

+++ devstack-vm-gate-wrap.sh
@@ -49,7 +49,10 @@ function git_checkout {
 
     git checkout $branch
     git reset --hard $reset_branch
-    git clean -x -f -d -q
+    if ! git clean -x -f -d -q ; then
+        sleep 1
+        git clean -x -f -d -q
+    fi
 }
 
 function fix_etc_hosts {
